// src/api/requests/swapi/homeworld/_types.ts
export type TWorld = {
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: Array<string>;
  created: string;
  edited: string;
  url: string;
};